// src/modules/activity-log/activity_log.enum.ts
export enum ActivityLogType {
  APPOINTMENT_CREATED = "APPOINTMENT_CREATED",
  APPOINTMENT_UPDATED = "APPOINTMENT_UPDATED",
  APPOINTMENT_CANCELLED = "APPOINTMENT_CANCELLED",
  APPOINTMENT_COMPLETED = "APPOINTMENT_COMPLETED",
  APPOINTMENT_NO_SHOW = "APPOINTMENT_NO_SHOW",
  STAFF_ASSIGNED = "STAFF_ASSIGNED",
  STAFF_UNASSIGNED = "STAFF_UNASSIGNED",
  QUEUE_TO_STAFF = "QUEUE_TO_STAFF",
  STAFF_TO_QUEUE = "STAFF_TO_QUEUE",
  APPOINTMENT_DELETED = "APPOINTMENT_DELETED",
}